feat: enhance tab and section management with import string support and refactor actions
