+++ src/components/shared/class-guides/editor/components/talents/schemas/talents-schemas.ts
@@ -1,7 +1,7 @@
 import { z } from 'zod';
 
 export const talentSchemas = z.object({
-  content: z.string(),
+  importString: z.string(),
 });
 
 export type TTalentSchemas = z.infer<typeof talentSchemas>;
