+++ src/components/shared/class-guides/editor/components/talents/talents-editor.tsx
@@ -10,15 +10,18 @@ import { FormInput } from '@root/components/shared/forms/input-form';
 import Image from 'next/image';
 import { useForm, FormProvider } from 'react-hook-form';
 import { cn } from '@root/lib/utils';
+import { MDTabContentEditor } from '../text-field/md-tab-content-editor';
 
 interface TalentsEditorProps {
   tab: TabData;
   specialization: ClassSpecialization;
   onContentChange: (value: string) => void;
+  onImportChange: (value: string) => void;
 }
 
 interface TalentFormData {
-  talentString: string;
+  importString: string;
+  content: string;
 }
 
 // Фиксированные размеры для контейнера и iframe
@@ -29,18 +32,21 @@ export const TalentsEditor = ({
   specialization,
   tab,
   onContentChange,
+  onImportChange,
 }: TalentsEditorProps) => {
   const methods = useForm<TalentFormData>({
     defaultValues: {
-      talentString: '',
+      importString: tab.importString || '',
+      content: tab.content || '',
     },
   });
 
   const { handleSubmit, reset } = methods;
 
   const onSubmit = (data: TalentFormData) => {
-    if (data.talentString.trim()) {
-      onContentChange(data.talentString);
+    if (data.importString.trim()) {
+      onContentChange(data.content);
+      onImportChange(data.importString);
       reset();
     }
   };
@@ -56,7 +62,7 @@ export const TalentsEditor = ({
             }}
           >
             {/* Фоновое изображение через Image */}
-            <div className='absolute inset-0 opacity-100 brightness-100'>
+            <div className='absolute inset-0 border-2 border-black opacity-100 brightness-100'>
               <Image
                 src={specialization.talentsBG}
                 alt='Talents background'
@@ -72,9 +78,9 @@ export const TalentsEditor = ({
                 style={{ colorScheme: 'light' }}
                 className='relative z-20 flex h-full items-center justify-center'
               >
-                {tab.content ? (
+                {tab.importString ? (
                   <iframe
-                    src={`https://www.raidbots.com/simbot/render/talents/${tab.content}?locale=ru_RU&width=${IFRAME_WIDTH}&level=80&bgcolor=transparent`}
+                    src={`https://www.raidbots.com/simbot/render/talents/${tab.importString}?locale=ru_RU&width=${IFRAME_WIDTH}&level=80&bgcolor=transparent`}
                     width={IFRAME_WIDTH}
                     height={IFRAME_HEIGHT}
                     title='Talents Editor'
@@ -85,14 +91,14 @@ export const TalentsEditor = ({
                   <FormProvider {...methods}>
                     <div className='flex flex-col items-center gap-4 pt-20'>
                       <FormInput
-                        name='talentString'
+                        name='importString'
                         placeholder='Введите строку талантов'
                         className='w-[600px]'
                         required
                       />
                       <Button
                         onClick={handleSubmit(onSubmit)}
-                        disabled={!methods.watch('talentString').trim()}
+                        disabled={!methods.watch('importString').trim()}
                       >
                         Отправить
                       </Button>
@@ -103,6 +109,11 @@ export const TalentsEditor = ({
               <ScrollBar orientation='horizontal' />
             </ScrollArea>
           </div>
+          <MDTabContentEditor
+            className='w-full px-2'
+            content={tab.content}
+            onContentChange={onContentChange}
+          />
         </CardContent>
       </Card>
     </TabsContent>
